+++ CHANGELOG.md
@@ -6,3 +6,11 @@ Initial Release
 
 ### Added:
 - export xlf tools
+
+## [1.1.0] - 2025-01-31
+
+### Added:
+- Little Endian Conversion
+
+### Breaking:
+- `xflToHex(value)` & `hexToXFl(value)` now return the Big Endian by default. To get the Little Endian you need to do: `xflToHex(value, true)` and `hexToXFl(value, true)`